Collapse composer license plugin
